Move ServerContext to server_context.py
