+++ labkey/container.py
@@ -1,4 +1,5 @@
-from labkey.utils import json_dumps, ServerContext
+from .server_context import ServerContext
+from labkey.utils import json_dumps
 
 
 def create(
